Update AboutFragment.kt
We added a variable called "timer", which is initialized upon the first click on the app_logo. The timer starts a countdown for 5 seconds, after which, if there have been at least 7 clicks, an animation is triggered. If there were less than 7 clicks, they are reset. If there were already more than 7 clicks, the timer is stopped and the animation is triggered. Additionally, we added the disabling of the ability to click on app_logo during the animation to avoid unwanted repeated clicks.